dev_team/requirements.txt
PyGithub
langchain
chromadb
openai
colorama
langgraph